Changed integer to decimal for harga

+++ JRTParcel/database/migrations/2024_06_19_134645_add_harga_to_resi_table.php
@@ -14,7 +14,7 @@ class AddHargaToResiTable extends Migration
     public function up()
 {
     Schema::table('resi', function (Blueprint $table) {
-        $table->integer('harga')->default(0)->after('kecamatan_kota_tujuan');
+        $table->decimal('harga',8,2)->default(0)->after('kecamatan_kota_tujuan');
     });
 }
